+++ .gitignore
@@ -1,1 +1,3 @@
 /.metadata/
+/Covid-19 Vaccination Registration System/nbproject/private/
+/Covid-19 Vaccination Registration System/build/
