--- Covid-19 Vaccination Registration System/src/clientPersonnel/PersonnelMainMenu.java
@@ -1,43 +1,0 @@
-package clientPersonnel;
-
-import java.awt.EventQueue;
-
-import javax.swing.JFrame;
-
-public class PersonnelMainMenu {
-
-	private JFrame frame;
-
-	/**
-	 * Launch the application.
-	 */
-	public static void main(String[] args) {
-		EventQueue.invokeLater(new Runnable() {
-			public void run() {
-				try {
-					PersonnelMainMenu window = new PersonnelMainMenu();
-					window.frame.setVisible(true);
-				} catch (Exception e) {
-					e.printStackTrace();
-				}
-			}
-		});
-	}
-
-	/**
-	 * Create the application.
-	 */
-	public PersonnelMainMenu() {
-		initialize();
-	}
-
-	/**
-	 * Initialize the contents of the frame.
-	 */
-	private void initialize() {
-		frame = new JFrame();
-		frame.setBounds(100, 100, 819, 482);
-		frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-	}
-
-}
